+++ api/axiosInstance.ts
@@ -1,0 +1,6 @@
+import axios from "axios";
+
+export const instance = axios.create({
+  baseURL: "https://bootcamp-api.codeit.kr/api/linkbrary/v1",
+  withCredentials: true,
+});
