+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@tanstack/react-query": "^5.36.0",
     "@tanstack/react-query-devtools": "^5.36.0",
+    "axios": "^1.6.8",
     "next": "^14.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
